// web/remote-control/src/generated/types/AuthChallenge.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ChallengeId } from "./ChallengeId";
import type { DeviceId } from "./DeviceId";

export type AuthChallenge = { id: ChallengeId, device_id: DeviceId, challenge: string, expires_at: string, };