// web/remote-control/src/generated/types/Authenticated.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AuthSessionId } from "./AuthSessionId";
import type { DeviceSummary } from "./DeviceSummary";

export type Authenticated = { session_id: AuthSessionId, device: DeviceSummary, expires_at: string, replayed_from_sequence: number | null, };