// web/remote-control/src/generated/types/PairingStatus.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Capability } from "./Capability";
import type { DeviceId } from "./DeviceId";

export type PairingStatus = { "status": "pending" } | { "status": "approved", "data": { device_id: DeviceId, capabilities: Array<Capability>, } } | { "status": "rejected" } | { "status": "expired" };